+++ docs/source/conf.py
@@ -11,8 +11,8 @@ sys.path.insert(0, os.path.abspath('../..')) #sphinx access
 project = 'SubCortexMesh'
 copyright = '2026, Cognitive and Brain Health Laboratory - Nanyang Technological University'
 author = 'Charly H. A. Billaud, Nicolas P.M. Lavarde, Junhong Yu'
-release = '1.1.0'
-version = '1.1.0'
+release = '1.2.0'
+version = '1.2.0'
 
 # -- General configuration ---------------------------------------------------
 extensions = [